cognitum/hf_space/requirements.txt
gradio==5.29.0
pyyaml
typer
